refactor(workspace): extract command palette selection hook
